made Panel a global component & refactored
