+++ client/src/components/Register.vue
@@ -24,7 +24,6 @@
 <script>
 import AuthenticationService from "@/services/AuthenticationService";
 import store from "../store";
-import Panel from "@/components/Panel.vue";
 
 export default {
     data() {
@@ -47,9 +46,6 @@ export default {
                 this.error = error.response.data.error;
             }
         },
-    },
-    components: {
-        Panel,
     },
 };
 </script>
