+++ client/src/components/Songs/SongsSearchPanel.vue
@@ -1,0 +1,9 @@
+<template>
+    <panel title="Search" />
+</template>
+
+<script>
+export default {};
+</script>
+
+<style scoped></style>
