+++ client/src/components/ViewSong/Lyrics.vue
@@ -5,11 +5,7 @@
 </template>
 
 <script>
-import Panel from "@/components/Panel.vue";
 export default {
-    components: {
-        Panel,
-    },
     props: ["song"],
 };
 </script>
