+++ client/src/components/ViewSong/SongMetadata.vue
@@ -35,11 +35,7 @@
 </template>
 
 <script>
-import Panel from "@/components/Panel.vue";
 export default {
-    components: {
-        Panel,
-    },
     props: ["song"],
     methods: {
         navigateTo(route) {
