+++ client/src/components/globals/Panel.vue
